fixup(services): complete world_entry_methods rename cleanup
Updates leftover use paths and removes the dangling
`pub(crate) mod world_entry_methods;` declaration in `base/mod.rs`,
finishing the rename `world_entry_methods/` -> `world_entry/methods/`
introduced in eb12758. Also renames the inner `world_entry.rs` to
`world_entry_db.rs` to avoid a name-collision with the parent module.
