+++ crates/services/src/base/world_entry/methods/mod.rs
@@ -2,7 +2,9 @@
 // All functions previously in world_entry_player.rs are now distributed here.
 // Organized into semantic groups: player_load/, inventory/, vendor/ for maintainability.
 
-pub mod world_entry;
+// Renamed from `world_entry.rs` to avoid name-collision with the parent
+// `world_entry/` module after the split-and-consolidate refactor.
+pub mod world_entry_db;
 pub mod progression;
 pub mod missions;
 pub mod mail;
@@ -11,7 +13,7 @@ pub mod inventory;
 pub mod vendor;
 
 // Re-export all public functions for backward compatibility
-pub use world_entry::{query_world_entry, query_world_stargates};
+pub use world_entry_db::{query_world_entry, query_world_stargates};
 pub use player_load::{query_player_load_data, query_player_load_data_by_account, query_inventory_items, default_player_load_data, query_bandolier_items, query_archetype_ability_tree};
 pub use progression::{handle_grant_xp, handle_grant_cash};
 pub use missions::{query_saved_missions, handle_mission_update};
